XFIRE-564: ClassCastException: org.codehaus.xfire.aegis.type.mtom.ByteArrayType expects the underlying DataSource to be an AttachmentDataSource
git-svn-id: http://svn.codehaus.org/xfire/trunk/xfire@1794 c7d0bf07-ec0d-0410-b2cc-d48fa9be22ba

+++ xfire-aegis/src/main/org/codehaus/xfire/aegis/type/mtom/ByteArrayType.java
@@ -7,6 +7,7 @@ import java.io.InputStream;
 import java.io.OutputStream;
 
 import javax.activation.DataHandler;
+import javax.activation.DataSource;
 import javax.xml.namespace.QName;
 
 import org.codehaus.xfire.MessageContext;
@@ -40,9 +41,12 @@ public class ByteArrayType extends AbstractXOPType
         }
         finally
         {
-            AttachmentDataSource dataSource = (AttachmentDataSource) handler.getDataSource();
-            File attFile = dataSource.getFile();
-            if (attFile != null) attFile.delete();
+            DataSource dataSource = handler.getDataSource();
+            if (dataSource instanceof AttachmentDataSource) 
+            {
+                File attFile = ((AttachmentDataSource) dataSource).getFile();
+                if (attFile != null) attFile.delete();
+            }
         }
     }
